requirements.txt
numpy
Pillow
tkinterdnd2